add colors to agenda, see description for more info
i modified agenda to have a cssName, which will be used in the jinja to generate a class in the html. That class has to be one of the classes inside css/agenda-accordion-colors. Also, i made the accordion fonts bigger and with more font weight because i personally had trouble visualizing them, so i imagine anyone else would have the same exact problem :). Hope you have a good day✨